feat:added plugins chips filter in microsite
